@@ -3,13 +3,11 @@ require 'spellparse'
 require 'htmlentities'
 
 NORMAL_SELECT = "DISTINCT spells.*, r1.name AS reagent1_name, r2.name AS reagent2_name, skill.name AS skill_name, target.name AS target_name, resist.name AS resist_name, recourse.name AS recourse_name"
-NORMAL_JOIN = "LEFT JOIN reagents AS r1 ON r1.id = spells.reagent1_id LEFT JOIN reagents as r2 ON r2.id = spells.reagent2_id INNER JOIN skill_types as skill ON skill.id = spells.skill_id "
-NORMAL_JOIN += "INNER JOIN target_types AS target ON target.id = spells.target_type_id LEFT JOIN resist_types AS resist ON resist.id = spells.resist_type_id "
-NORMAL_JOIN += "LEFT JOIN spells AS recourse ON recourse.id = spells.recourse_id "
+NORMAL_JOIN = "LEFT JOIN reagents AS r1 ON r1.id = spells.reagent1_id LEFT JOIN reagents as r2 ON r2.id = spells.reagent2_id INNER JOIN skill_types as skill ON skill.id = spells.skill_id INNER JOIN target_types AS target ON target.id = spells.target_type_id LEFT JOIN resist_types AS resist ON resist.id = spells.resist_type_id LEFT JOIN spells AS recourse ON recourse.id = spells.recourse_id"
 
 class SearchSpellsController < ApplicationController
 
-	before_filter :sanitize_params
+	before_action :sanitize_params
 
 protected
 	def valid_int?(i)
@@ -214,7 +212,7 @@ public
 		# For a given spell, pull in its classes
 		@spells.each do | spell |
 			@class = GetSpellClasses(spell.id)
-			spell["classes"] = @class.first["classes"]
+			spell.classes = @class.first["classes"]
 
 			# Also pull in effects and calculate results
 			@arrEffects,@hashStacks = GetSpellEffects(spell.id, spell.beneficial, spell.duration, spell.extra)
@@ -294,8 +292,9 @@ public
 			
 			# For a given spell, pull in its classes
 			@spells.each do | spell |
-				@class = MapSpellToCharClass.find_by_sql("SELECT group_concat(CONCAT(char_classes.name, '(', m.level, ')') SEPARATOR ', ') AS classes FROM char_classes INNER JOIN map_spell_to_char_classes as m ON m.class_id = char_classes.id WHERE m.spell_id = " + spell["id"].to_s + " ORDER BY level, name")
-				spell["classes"] = @class.first["classes"]
+				#@class = MapSpellToCharClass.find_by_sql("SELECT group_concat(CONCAT(char_classes.name, '(', m.level, ')') SEPARATOR ', ') AS classes FROM char_classes INNER JOIN map_spell_to_char_classes as m ON m.class_id = char_classes.id WHERE m.spell_id = " + spell["id"].to_s + " ORDER BY level, name")
+				@class = GetSpellClasses(spell.id)
+				spell.classes = @class.first["classes"]
 			end
 
 			if (@spells.length == 1)
